Compiler passes merged into one
