+++ DizdaCloudBackupBundle.php
@@ -2,9 +2,7 @@
 
 namespace Dizda\CloudBackupBundle;
 
-use Dizda\CloudBackupBundle\DependencyInjection\Compiler\ClientCompilerPass;
-use Dizda\CloudBackupBundle\DependencyInjection\Compiler\DatabaseCompilerPass;
-use Dizda\CloudBackupBundle\DependencyInjection\Compiler\ProcessorCompilerPass;
+use Dizda\CloudBackupBundle\DependencyInjection\Compiler\TaggedServicesPass;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
@@ -16,8 +14,6 @@ class DizdaCloudBackupBundle extends Bundle
      */
     public function build(ContainerBuilder $container)
     {
-        $container->addCompilerPass(new DatabaseCompilerPass());
-        $container->addCompilerPass(new ClientCompilerPass());
-        $container->addCompilerPass(new ProcessorCompilerPass());
+        $container->addCompilerPass(new TaggedServicesPass());
     }
 }
